+++ client/src/views/home/Home.jsx
@@ -1,15 +1,20 @@
-import React from "react";
+import React, { useEffect } from "react";
+import { useDispatch } from "react-redux";
 import Features from "../../components/Features/Features";
 import Footer from "../../components/Footer/Footer";
 import Header from "../../components/Header/Header";
-import Hero from "../../components/Hero/Hero";
 import HeroMain from "../../components/HeroMain/HeroMain";
-import MenuHome from "../../components/MenuHome/MenuHome";
-import Opiniones from "../../components/Opiniones/Opiniones";
 import OpinionesContainer from "../../components/OpinionesContainer/OpinionesContainer";
+import { getReview } from "../../redux/actions/actions";
 import SliderSmall from "../../components/SliderSmall/SliderSmall";
 
 export default function Home() {
+	const dispatch = useDispatch();
+
+	useEffect(() => {
+		dispatch(getReview());
+	}, [dispatch]);
+
 	return (
 		<div>
 			<Header />
